ci: use Build Stages on Travis CI
Update the configuration for Travis CI to use build stages, allowing a
stable and beta build of Atom to run alongside a dedicated linting job
that checks the commit message and runs the `lint` script.

The deploy stage will only run for commits directly to `master` now.

The build will also automatically use the current LTS release of
Node.js.